+++ postcss.config.js
@@ -1,6 +1,5 @@
 import postcssImport from 'postcss-import'
 import postcssPresetEnv from 'postcss-preset-env'
-import autoprefixer from 'autoprefixer'
 import tailwindcss from 'tailwindcss'
 
 // https://github.com/postcss/postcss#usage
@@ -9,8 +8,9 @@ import tailwindcss from 'tailwindcss'
 export default {
   plugins: [
     postcssImport(),
-    postcssPresetEnv(),
-    autoprefixer(),
     tailwindcss(),
+    postcssPresetEnv({
+      stage: 1,
+    }),
   ],
 }